replaced account interface with entity in service and controller; removed interface
